fix: 严格 c++17 下 glibc 宏兼容（EPOLLNVAL→POLLNVAL）
- Channel：glibc <sys/epoll.h> 不提供 EPOLLNVAL，改用 <poll.h> 的 POLLNVAL
- EPollPoller：修 -Wsign-compare（int vs size_t 比较）

+++ src/EPollPoller.cc
@@ -32,7 +32,7 @@ Timestamp EPollPoller::poll(int timeoutMs, Poller::ChannelList* activeChannels)
 	if (nfds > 0)
 	{
 		fillActiveChannels(nfds, activeChannels);
-		if (static_cast<int>(nfds) == events_.size())
+		if (static_cast<size_t>(nfds) == events_.size())
 		{
 			events_.resize(events_.size() * 2);
 		}
